docker/services/image/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
transformers==4.45.2
diffusers==0.31.0
accelerate==1.1.1
safetensors==0.4.5
torchvision==0.19.0
pillow==10.4.0
numpy==1.26.4
